requirements.txt
# 阅卷系统依赖包

# 浏览器自动化
playwright>=1.41.0

# OCR 引擎
paddleocr>=2.7.0
paddlepaddle>=2.6.0

# AI 调用
openai>=1.10.0

# 图像处理
pillow>=10.2.0

# 配置文件解析
pyyaml>=6.0.1

# 打包工具
pyinstaller>=6.4.0

# 数值计算
numpy>=1.26.3

# 加密
cryptography>=41.0.0
